requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pyyaml==6.0.1

# Data Processing
numpy==1.26.3
pandas==2.1.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
hypothesis==6.92.2

# Development
black==23.12.1
ruff==0.1.11
mypy==1.8.0
